Created a test button for user input
Created a TEST button to read input as a variable for the number of sectionals.

This was not ran beforehand due to issues with Visual Studio Code, but it did compile successfully.

Will need to see if someone else can run it and go from there: Implement for all buttons we have and then use the sectional variable "ad" for the algorithm runs.